refactor: rename api/services/settings → api/services/configuration (CHAOS-1559)
